feat|fix: adding attribute to skip null properties
Adding the [SerializeIfNotNullAttribute] which can be applied to
properties serialized by the QuerySerializer to indicate they should
never be serialized when null.

As a proof of concept, the RuleParameterInput used in the creation of
rulesets for the repository was failing due to null properties being
seralized into the final output. This code now functions as expected.

I did not make a global change to no longer serialized all null values
as I anticipate that there are "side effects" we depend on where there
are expected null serialzations that will break if such a change is
made.

This is an "opt-in" feature that does not disturb the greater echosystem
that may depend on these "side effects".

octokit/octokit.graphql.net#320